- add CTRL-I and CTRL-D shortcuts for initializing and deinitializing applications - call detach() slots of services before removing them (only if present), use them in the GuiLogger service - fix deprectation warning about float vs int in BrowserWidget - add a test case for the deadlocking use case

+++ nexxT/services/gui/BrowserWidget.py
@@ -166,7 +166,7 @@ class FolderListModel(QAbstractTableModel):
                 return s
             if index.column() == 2:
                 try:
-                    return QDateTime.fromMSecsSinceEpoch(c.stat().st_mtime*1000)
+                    return QDateTime.fromMSecsSinceEpoch(int(c.stat().st_mtime*1000))
                 except Exception: # pylint: disable=broad-except
                     return ""
         if role == Qt.DecorationRole:
